Add authentication middleware and route flow
