+++ middleware.ts
@@ -1,0 +1,23 @@
+import { NextResponse } from "next/server";
+import type { NextRequest } from "next/server";
+
+export function middleware(req: NextRequest) {
+  const token = req.cookies.get("token")?.value;
+  const { pathname } = req.nextUrl;
+
+  const isLoginRoute = pathname === "/login";
+
+  if (!token && !isLoginRoute) {
+    return NextResponse.redirect(new URL("/login", req.url));
+  }
+
+  if (token && isLoginRoute) {
+    return NextResponse.redirect(new URL("/", req.url));
+  }
+
+  return NextResponse.next();
+}
+
+export const config = {
+  matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
+};
